src/requirements.txt
ansicon==1.89.0
APScheduler==3.10.1
arrow==1.2.3
asgiref==3.6.0
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-environ==0.10.0
django-filter==22.1
django-model-utils==4.3.1
django-phonenumber-field==7.1.0
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
jinxed==1.2.0
Markdown==3.4.1
mysqlclient==2.1.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
wcwidth==0.2.6
whitenoise==6.4.0
django-crontab==0.7.1